boot: move Initramfs*Dir vars to separate file
